COMMIT Mudanças Finais e Polimento
Polimento do código, algumas trocas de if e else if por switch
Limpeza e tratamento de estética dos métodos
Troca do tipo INT para LONG do atributo cpf
Adição do método virtual void get_tipo() -> Filme.hpp, dvd.hpp,
                                            dvd.cpp, fita.cpp, fita.hpp
Log funcionando, de maneira que é truncado (substitui o último log)
toda vez que se inicia o programa;
Vários testes manuais realizados, mas está na hora de fazer os outros.

+++ Locadora.hpp
@@ -29,13 +29,13 @@ class Locadora
 
         void getNome();
 
-        void alugar_Filme( vector<int> codigos, int cpf );
-        void devolver_Filme( int cpf );
+        void alugar_Filme( vector<int> codigos, long cpf );
+        void devolver_Filme( long cpf );
 
-        void cadastrar_cliente( int cpf, string nome);
-        void remover_cliente(int cpf);
+        void cadastrar_cliente( long cpf, string nome);
+        void remover_cliente(long cpf);
         void imprimir_clientes(char tipo_ordenacao);
-        Cliente* buscar_cliente(int cpf);
+        Cliente* buscar_cliente(long cpf);
 
         map <int, Filme*> getEstoque();
